correct metric name
